\name{randomTrunc-package}
\alias{randomTrunc-package}
\alias{randomTrunc}
\docType{package}
\title{
\packageTitle{randomTrunc}
}
\description{
\packageDescription{randomTrunc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{randomTrunc}
\packageIndices{randomTrunc}
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
Generating the density and random generation for the truncated normal distribution with a truncation point or a normal distributed truncation threshold.
}
\author{
\packageAuthor{randomTrunc}

Maintainer: \packageMaintainer{randomTrunc}
}
% \references{
% ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
\keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
